Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
requests
duckduckgo-search
chromadb
sentence-transformers
pydantic